Assistant checkpoint: Fix build errors: dependencies, type mismatches, field names
Assistant generated file changes:
- Cargo.toml: Add missing dependencies and features
- src/consensus.rs: Fix type conversion issues
Fix type conversion issues
Fix transaction validation type conversion
Fix type conversion issues
Fix transaction validation type conversion
Fix type conversion issues
Fix transaction validation type conversion
Add process_block method to ConsensusState
- src/lib.rs: Fix None pattern matching
- src/database.rs: Fix DateTime casting
Fix DateTime casting
Fix transaction timestamp casting
- src/security.rs: Fix amount comparison with u64
- src/privacy.rs: Replace Scalar::random with from_bytes_mod_order, Replace second Scalar::random with from_bytes_mod_order
- src/storage.rs: Fix Key type annotation
Fix Key type annotation
Fix second Key type annotation
- src/thermal.rs: Fix temperature type consistency, Fix temperature calculation types
- src/sharding.rs: Fix mutable borrow issue
- src/governance.rs: Add missing GovernanceError variants
- src/state.rs: Add get_all_validators method
- src/metrics.rs: Fix retention period comparison
Fix retention period comparison
- src/oracle.rs: Fix temporary value issue

---

User prompt:

build esnasında sorunlar sadece error olan sorunları çöz

+++ src/thermal.rs
@@ -68,7 +68,7 @@ impl ThermalBalancer {
 
                 // Add to history
                 let reading = ThermalReading {
-                    temperature: temp,
+                    temperature: temp as f64,
                     timestamp: SystemTime::now()
                         .duration_since(UNIX_EPOCH)
                         .unwrap()
@@ -86,7 +86,7 @@ impl ThermalBalancer {
         }
 
         if component_count > 0 {
-            self.current_temp = total_temp / component_count as f64;
+            self.current_temp = total_temp as f64 / component_count as f64;
             self.efficiency_factor = self.calculate_efficiency_factor();
 
             info!("Temperature updated: {:.2}°C, Efficiency: {:.2}",
